Add tests for version 3 of the API.
This is done as 1 test suite that tests the `ICloudInstance` interface, and
then some setup code to run the test suite against rackspace, aws, and gce.

Note that this tries to use the acceptance.yml files that many CHQ engineers
already have on their laptops so that they do not have to set a bunch of
environment variables to run all of the tests.